# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mhauri AI Agriculture Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export interface ConversationInput {
  /** @minLength 1 */
  title: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export type ArticleLanguage = typeof ArticleLanguage[keyof typeof ArticleLanguage];


export const ArticleLanguage = {
  en: 'en',
  sn: 'sn',
  nd: 'nd',
  all: 'all',
} as const;

export interface Article {
  id: number;
  title: string;
  content: string;
  category: string;
  tags: string[];
  language: ArticleLanguage;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export type ArticleInputLanguage = typeof ArticleInputLanguage[keyof typeof ArticleInputLanguage];


export const ArticleInputLanguage = {
  en: 'en',
  sn: 'sn',
  nd: 'nd',
  all: 'all',
} as const;

export interface ArticleInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  /** @minLength 1 */
  category: string;
  tags?: string[];
  language?: ArticleInputLanguage;
  isActive?: boolean;
}

export type ArticleUpdateLanguage = typeof ArticleUpdateLanguage[keyof typeof ArticleUpdateLanguage];


export const ArticleUpdateLanguage = {
  en: 'en',
  sn: 'sn',
  nd: 'nd',
  all: 'all',
} as const;

export interface ArticleUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  category?: string;
  tags?: string[];
  language?: ArticleUpdateLanguage;
  isActive?: boolean;
}

export type FarmerLanguagePref = typeof FarmerLanguagePref[keyof typeof FarmerLanguagePref];


export const FarmerLanguagePref = {
  en: 'en',
  sn: 'sn',
  nd: 'nd',
} as const;

export interface Farmer {
  phone: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  location?: string | null;
  crops: string[];
  livestock: string[];
  languagePref: FarmerLanguagePref;
  isActive: boolean;
  lastSeen: string;
  createdAt: string;
  updatedAt: string;
}

export type FarmerUpdateLanguagePref = typeof FarmerUpdateLanguagePref[keyof typeof FarmerUpdateLanguagePref];


export const FarmerUpdateLanguagePref = {
  en: 'en',
  sn: 'sn',
  nd: 'nd',
} as const;

export interface FarmerUpdate {
  name?: string;
  location?: string;
  crops?: string[];
  livestock?: string[];
  languagePref?: FarmerUpdateLanguagePref;
  isActive?: boolean;
}

export interface MarketPrice {
  id: number;
  commodity: string;
  unit: string;
  priceUsd: string;
  market: string;
  priceDate: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface MarketPriceInput {
  /** @minLength 1 */
  commodity: string;
  unit?: string;
  priceUsd: string;
  market?: string;
  priceDate: string;
  notes?: string;
}

export type BroadcastStatus = typeof BroadcastStatus[keyof typeof BroadcastStatus];


export const BroadcastStatus = {
  draft: 'draft',
  sending: 'sending',
  sent: 'sent',
  failed: 'failed',
} as const;

export interface Broadcast {
  id: number;
  message: string;
  status: BroadcastStatus;
  recipientCount: number;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
}

export interface BroadcastInput {
  /** @minLength 1 */
  message: string;
}

export type AnalyticsSummaryLanguageBreakdown = {[key: string]: number};

export type AnalyticsSummaryMessagesPerDayItem = {
  date: string;
  count: number;
};

export type AnalyticsSummaryTopEventTypesItem = {
  eventType: string;
  count: number;
};

export interface AnalyticsSummary {
  totalMessages: number;
  totalFarmers: number;
  languageBreakdown: AnalyticsSummaryLanguageBreakdown;
  messagesPerDay: AnalyticsSummaryMessagesPerDayItem[];
  topEventTypes: AnalyticsSummaryTopEventTypesItem[];
}

export type ListArticlesParams = {
category?: string;
language?: string;
search?: string;
};

export type ListFarmersParams = {
search?: string;
isActive?: boolean;
};

export type ListMarketPricesParams = {
commodity?: string;
market?: string;
};

export type GetAnalyticsSummaryParams = {
days?: number;
};