# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mhauri AI Agriculture Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsSummaryLanguageBreakdown } from './analyticsSummaryLanguageBreakdown';
import type { AnalyticsSummaryMessagesPerDayItem } from './analyticsSummaryMessagesPerDayItem';
import type { AnalyticsSummaryTopEventTypesItem } from './analyticsSummaryTopEventTypesItem';

export interface AnalyticsSummary {
  totalMessages: number;
  totalFarmers: number;
  languageBreakdown: AnalyticsSummaryLanguageBreakdown;
  messagesPerDay: AnalyticsSummaryMessagesPerDayItem[];
  topEventTypes: AnalyticsSummaryTopEventTypesItem[];
}