# lib/api-zod/src/generated/types/broadcast.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mhauri AI Agriculture Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { BroadcastStatus } from './broadcastStatus';

export interface Broadcast {
  id: number;
  message: string;
  status: BroadcastStatus;
  recipientCount: number;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
}