# lib/api-zod/src/generated/types/broadcastInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mhauri AI Agriculture Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface BroadcastInput {
  /** @minLength 1 */
  message: string;
}