# lib/api-zod/src/generated/types/marketPrice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mhauri AI Agriculture Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketPrice {
  id: number;
  commodity: string;
  unit: string;
  priceUsd: string;
  market: string;
  priceDate: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}